refactor: standardize gRPC error handling, expand domain request models, and migrate Kubernetes manifests to kustomize-based architecture.

+++ merchant_detail/src/main/java/io/example/merchant_detail/repository/MerchantSocialLinkCommandRepository.java
@@ -1,16 +1,22 @@
 package io.example.merchant_detail.repository;
 
+import io.example.merchant_detail.domain.requests.CreateMerchantSocialRequest;
+import io.example.merchant_detail.domain.requests.UpdateMerchantSocialRequest;
 import io.example.merchant_detail.model.MerchantSocialMediaLink;
 import io.vertx.core.Future;
-import pb.MerchantSocialLinkCommand.CreateMerchantSocialRequest;
-import pb.MerchantSocialLinkCommand.UpdateMerchantSocialRequest;
 
 public interface MerchantSocialLinkCommandRepository {
   Future<MerchantSocialMediaLink> create(CreateMerchantSocialRequest req);
+
   Future<MerchantSocialMediaLink> update(UpdateMerchantSocialRequest req);
-  Future<MerchantSocialMediaLink> trash(Long id);
-  Future<MerchantSocialMediaLink> restore(Long id);
-  Future<Void> deletePermanent(Long id);
+
+  Future<MerchantSocialMediaLink> trash(Integer id);
+
+  Future<MerchantSocialMediaLink> restore(Integer id);
+
+  Future<Void> deletePermanent(Integer id);
+
   Future<Integer> restoreAll();
+
   Future<Integer> deleteAll();
 }
