+++ order/src/main/java/io/example/order/repository/TransactionCommandRepository.java
@@ -3,6 +3,6 @@ package io.example.order.repository;
 import io.vertx.core.Future;
 
 public interface TransactionCommandRepository {
-    Future<Boolean> deleteByOrderIDPermanent(Integer orderId);
+    Future<Boolean> deleteByOrderIDPermanent(Long orderId);
     Future<Boolean> deleteAll();
 }
